[SDK] refactor code style
