+++ inlong-sdk/dataproxy-sdk/src/main/java/org/apache/inlong/sdk/dataproxy/common/FileCallback.java
@@ -15,22 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sdk.dataproxy;
+package org.apache.inlong.sdk.dataproxy.common;
 
-public enum SendResult {
-    INVALID_ATTRIBUTES, // including DataProxyErrCode(100,101,102,112)
-    OK,
-    TIMEOUT,
-    CONNECTION_BREAK,
-    THREAD_INTERRUPT,
-    ASYNC_CALLBACK_BUFFER_FULL,
-    NO_CONNECTION,
-    INVALID_DATA, // including DataProxyErrCode(103, 111)
-    BODY_EXCEED_MAX_LEN, // DataProxyErrCode(104)
-    SINK_SERVICE_UNREADY, // DataProxyErrCode(1)
-    UNCONFIGURED_GROUPID_OR_STREAMID, // DataProxyErrCode(113)
-    TOPIC_IS_BLANK, // DataProxyErrCode(115)
-    DATAPROXY_FAIL_TO_RECEIVE, // DataProxyErrCode(114,116,117,118,119,120)
+public abstract class FileCallback implements SendMessageCallback {
 
-    UNKOWN_ERROR
+    /* Invoked when a message is confirmed by TDBus. */
+    public void onMessageAck(String result) {
+    }
+
+    ;
+
+    public void onMessageAck(SendResult result) {
+    }
+
+    ;
+
+    /* Invoked when a message transportation interrupted by an exception. */
+    public void onException(Throwable e) {
+    }
+
+    ;
 }
