+++ inlong-sdk/dataproxy-sdk/src/main/java/org/apache/inlong/sdk/dataproxy/common/SendMessageCallback.java
@@ -15,24 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sdk.dataproxy;
+package org.apache.inlong.sdk.dataproxy.common;
 
-public abstract class FileCallback implements SendMessageCallback {
+public interface SendMessageCallback {
 
     /* Invoked when a message is confirmed by TDBus. */
-    public void onMessageAck(String result) {
-    }
-
-    ;
-
-    public void onMessageAck(SendResult result) {
-    }
-
-    ;
+    public void onMessageAck(SendResult result);
 
     /* Invoked when a message transportation interrupted by an exception. */
-    public void onException(Throwable e) {
-    }
-
-    ;
+    public void onException(Throwable e);
 }
