+++ inlong-sdk/dataproxy-sdk/src/main/java/org/apache/inlong/sdk/dataproxy/pb/PbProtocolMessageSender.java
@@ -20,8 +20,8 @@ package org.apache.inlong.sdk.dataproxy.pb;
 import org.apache.inlong.common.util.NetworkUtils;
 import org.apache.inlong.sdk.commons.protocol.SdkEvent;
 import org.apache.inlong.sdk.dataproxy.MessageSender;
-import org.apache.inlong.sdk.dataproxy.SendMessageCallback;
-import org.apache.inlong.sdk.dataproxy.SendResult;
+import org.apache.inlong.sdk.dataproxy.common.SendMessageCallback;
+import org.apache.inlong.sdk.dataproxy.common.SendResult;
 import org.apache.inlong.sdk.dataproxy.network.ProxysdkException;
 import org.apache.inlong.sdk.dataproxy.pb.channel.BufferQueueChannel;
 import org.apache.inlong.sdk.dataproxy.pb.context.CallbackProfile;
